feat: add base ci workflow (#2)
* feat: add base ci workflow
* feat: add clients exports and top level exports
* fix: simplify top level command exports
* fix: fix e2e imports
