+++ src/index.ts
@@ -1,0 +1,9 @@
+
+export { PollingStrategy } from "./wire/message/poll.utils.js";
+export { ConsumerKind } from "./wire/offset/offset.utils.js";
+export { Partitioning } from './wire/message/partitioning.utils.js';
+export { HeaderValue } from './wire/message/header.utils.js';
+
+export * from './client/index.js'
+export { ClientProvider, RawClient, CommandResponse } from './client/client.type.js';
+export { Id } from './wire/identifier.utils.js'
