+++ src/wire/client/index.ts
@@ -1,0 +1,4 @@
+
+export * from './get-client.command.js';
+export * from './get-clients.command.js';
+export * from './get-me.command.js';
