+++ src/wire/consumer-group/index.ts
@@ -1,0 +1,8 @@
+
+export * from './create-group.command.js';
+export * from './delete-group.command.js';
+export * from './get-group.command.js';
+export * from './get-groups.command.js';
+export * from './join-group.command.js';
+export * from './leave-group.command.js';
+
