+++ src/wire/index.ts
@@ -1,0 +1,12 @@
+
+export * from './client/index.js';
+export * from './consumer-group/index.js';
+export * from './message/index.js';
+export * from './offset/index.js';
+export * from './partition/index.js';
+export * from './session/index.js';
+export * from './stream/index.js';
+export * from './system/index.js';
+export * from './token/index.js';
+export * from './topic/index.js';
+export * from './user/index.js';
