+++ src/wire/offset/index.ts
@@ -1,0 +1,5 @@
+
+export * from './get-offset.command.js';
+export * from './store-offset.command.js';
+
+export { ConsumerKind } from './offset.utils.js';
