+++ src/wire/partition/index.ts
@@ -1,0 +1,3 @@
+
+export * from './create-partition.command.js';
+export * from './delete-partition.command.js';
