+++ src/wire/session/login.type.ts
@@ -1,0 +1,4 @@
+
+export type LoginResponse = {
+  userId: number
+}
