+++ src/wire/system/index.ts
@@ -1,0 +1,3 @@
+
+export * from './get-stats.command.js';
+export * from './ping.command.js';
