+++ src/wire/topic/index.ts
@@ -1,0 +1,7 @@
+
+export * from './create-topic.command.js';
+export * from './delete-topic.command.js';
+export * from './get-topic.command.js';
+export * from './get-topics.command.js';
+export * from './purge-topic.command.js';
+export * from './update-topic.command.js';
