+++ src/wire/user/index.ts
@@ -1,0 +1,8 @@
+
+export * from './change-password.command.js';
+export * from './create-user.command.js'
+export * from './delete-user.command.js';
+export * from './get-user.command.js'
+export * from './get-users.command.js';
+export * from './update-permissions.command.js';
+export * from './update-user.command.js';
