+++ tsconfig.json
@@ -10,6 +10,9 @@
     "forceConsistentCasingInFileNames": true,
     "moduleResolution": "node16",
 
+    "declaration": true,
+    "declarationMap": true,
+    
     "baseUrl": ".",
     "outDir": "./dist",
     "sourceMap": true
